Website

+++ Modules/Restaurant/database/migrations/2025_07_25_230359_add_restaurant_menu_category_id_to_restaurant_menu_items_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasColumn('restaurant_menu_items', 'restaurant_menu_categories_id')) {
+            return;
+        }
+
         Schema::table('restaurant_menu_items', function (Blueprint $table) {
             $table->foreignId('restaurant_menu_categories_id')->constrained('restaurant_menu_categories')->onDelete('cascade');
         });
@@ -22,7 +26,8 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('restaurant_menu_items', function (Blueprint $table) {
-            
+            $table->dropForeign(['restaurant_menu_categories_id']);
+            $table->dropColumn('restaurant_menu_categories_id');
         });
     }
 };
